+++ odoo/local-src/sf_account_financial_reporting/__manifest__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# Part of senseFly.
+
+{
+    'name': 'SenseFly/OCA account financial reporting',
+    'version': '1.0',
+    'author': 'Telmo Santos',
+    'category': 'SenseFly',
+    'depends': [
+        'account_financial_report_qweb'
+    ],
+    'data': [
+        'wizard/general_ledger_wizard_view.xml'
+    ],
+    'demo': [],
+    'images': ['static/description/icon.png'],
+    'installable': True,
+    'auto_install': False,
+}